// src/generated/soniox/schema/transcriptionStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Soniox Public API
 * OpenAPI spec version: 1.0.0
 */

export type TranscriptionStatus = typeof TranscriptionStatus[keyof typeof TranscriptionStatus];


export const TranscriptionStatus = {
  queued: 'queued',
  processing: 'processing',
  completed: 'completed',
  error: 'error',
} as const;